move unescape_error_reporting to lexer/
